The beginnings of the Faxanatools for Faxanadu.
Faxanadu is a criminally under-appreciated game for the NES, combining
action, platforming, and some puzzle-solving with an atmosphere you
don't often see on the NES.

I've been working on a project for doing a full disassembly of the game,
and have found some interesting elements, including level select code
and a full interaction scripting language. In order to better work on
this project and to play around with what I've learned, I built
Faxanatools.

Faxanatools is a Python package and command line tool for dumping content
from Faxanadu ROMs, modifying the content, and injecting back in. It can
be used today to customize some aspects of the game, and the hope is that,
down the road, it can be used to craft whole new adventures for our little
Faxanadu hero.

This is the initial commit for the project, based on what I put together
the past two nights. It can dump messages, interaction scripts, and shop
data from a Faxanadu ROM to a JSON format. Those can then be modified
and patched back into a ROM. Entire new NPC interactions can be written,
new shops added.

This still needs documentation on the file format and how the various
scripting actions and representations work, but as this is still pretty
much in flux, that will have to come at a much later date.